first service at works

+++ src/main/java/ikriti/natgeo/constants/MemberStatus.java
@@ -2,6 +2,6 @@ package ikriti.natgeo.constants;
 
 public class MemberStatus
 {
-	public static final int ACTIVE = 1;
-	public static final int INACTIVE = 2;
+	public static final Short ACTIVE = 1;
+	public static final Short INACTIVE = 2;
 }
